Combine the two device model files
Change-type: patch
