--- lib/models/device-ts.ts
@@ -1,11 +1,0 @@
-export enum OverallStatus {
-	CONFIGURING = 'configuring',
-	IDLE = 'idle',
-	OFFLINE = 'offline',
-	INACTIVE = 'inactive',
-	POST_PROVISIONING = 'post-provisioning',
-	UPDATING = 'updating',
-	ORDERED = 'ordered',
-	PREPARING = 'preparing',
-	SHIPPED = 'shipped',
-}
